Fixed this cursed problem
